Task 0

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -6,7 +6,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int nodes = 0;
+	size_t nodes = 0;
 
 	while (h != NULL)
 	{
